Remove presence_of validation for :credit_accepted.  Very strange behavior with shoulda_matchers. inclusion_of test gives a warning that false positives could potentially happen with 1s and 0s and possibly other values, but the tests all pass.
After the warning I changed the test to the more simple presence_of and spent a LOT of time (unproductive struggle) trying to troubleshoot to see if I had an accidental keystroke or syntax is just off, not sure. I have a feeling it will take very little to make the 'presence' test pass, but I don't know what it is, and it's apparently vital enough to throw an error and stop execution.